+++ src/lib/supabase/supabase.ts
@@ -9,46 +9,4 @@ if (!supabaseUrl || !supabaseAnonKey) {
 
 const supabase = createClient(supabaseUrl, supabaseAnonKey);
 
-// 구글 로그인 기능
-export const signInWithGoogle = async () => {
-  try {
-    const { data, error } = await supabase.auth.signInWithOAuth({
-      provider: "google",
-      options: {
-        queryParams: {
-          access_type: "offline",
-          prompt: "consent",
-        },
-      },
-    });
-
-    if (error) {
-      throw error;
-    }
-
-    return data;
-  } catch (error) {
-    throw error;
-  }
-};
-
-
-
-// 카카오 로그인 기능
-export const signInWithKakao = async () => {
-  try {
-    const { data, error } = await supabase.auth.signInWithOAuth({
-      provider: "kakao",
-    });
-
-    if (error) {
-      throw error;
-    }
-
-    return data;
-  } catch (error) {
-    throw error;
-  }
-};
-
 export default supabase;
